requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.2.3
numpy==1.26.4
matplotlib==3.8.2
seaborn==0.13.2
mysql-connector-python==9.2.0